% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian.R
\name{amplitudeResponse}
\alias{amplitudeResponse}
\title{Calculate the amplitude response based on R and Psi}
\usage{
amplitudeResponse(R, Psi, params)
}
\arguments{
\item{R}{Numeric, the value of R}

\item{Psi}{Numeric, the value of Psi}

\item{params}{List, a list of parameters including:
\itemize{
\item A_1, A_2: coefficients for the first and second terms
\item beta_L1, beta_L2: phase shifts for the first and second terms
}}
}
\value{
Numeric, the calculated amplitude response
}
\description{
Calculate the amplitude response based on R and Psi
}
