% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian.R
\name{circadianModel}
\alias{circadianModel}
\title{Calculate dR, dPsi, and dN based on the Hannay 19 et al. model}
\usage{
circadianModel(t, x, light, params)
}
\arguments{
\item{t}{Time in hours since the start of the week (0 = Sunday at midnight)}

\item{x}{Current values of the circadian model variables: R, Psi, and n.}

\item{light}{The light input to the circadian model.}

\item{params}{A list of parameters used in the circadian model.}
}
\value{
Derivatives of the circadian model variables: dR, dPsi, and dN.
}
\description{
Calculate dR, dPsi, and dN based on the Hannay 19 et al. model
}
