% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultParameters.R
\docType{data}
\name{defaultParameters}
\alias{defaultParameters}
\title{Default parameters for the several models and functions in the package}
\format{
A named vector with the following parameters
\describe{
  \item{A_1}{Circadian parameter}
  \item{A_2}{Circadian parameter}
  \item{sigma}{Circadian parameter}
  \item{beta_L1}{Circadian parameter}
  \item{beta_L2}{Circadian parameter}
  \item{P}{Circadian parameter}
  \item{G}{Circadian parameter}
  \item{I_0}{Circadian parameter}
  \item{delta}{Circadian parameter}
  \item{alpha_0}{Circadian parameter}
  \item{K}{Circadian parameter}
  \item{tau}{Circadian parameter}
  \item{gamma}{Circadian parameter}
  \item{beta_1}{Circadian parameter}
  \item{awakeMu}{Sleep parameter}
  \item{awakeChi}{Sleep parameter}
  \item{asleepMu}{Sleep parameter}
  \item{asleepChi}{Sleep parameter}
  \item{Irecep}{Sleep parameter}
  \item{targc}{Sleep parameter}
  \item{wakeThreshold}{Two process parameter}
  \item{sleepThreshold}{Two process parameter}
  \item{sleepDriveSlope}{Two process parameter}
  \item{schoolStart}{School parameter}
  \item{schoolDuration}{School parameter}
  \item{baselineLight}{Light parameter}
  \item{schoolBrightness}{Light parameter}
  \item{lightSlope}{Light parameter}
  \item{lightRiseTime}{Light parameter}
  \item{lightFallTime}{Light parameter}
  \item{lightAmplitude}{Light parameter}
  \item{lightBaseline}{Light parameter}
}
}
\source{
Arcascope
}
\usage{
defaultParameters
}
\description{
The named vector contains parameters for the circadian, sleep, and two
process models plus parameters for the light functions.
}
\details{
- 'A_1', 'A_2', 'sigma', 'beta_L1', 'beta_L2', 'P', 'G', 'I_0', 'delta', 'alpha_0', 'K', 'tau', 'gamma', 'beta_1': Forger99 variables for light ODEs (see reference below)
  *Forger, D. B., Jewett, M. E., & Kronauer, R. E. (1999). A simpler model of the human circadian pacemaker. Journal of 
       Biological Rhythms, 14(6), 533–538. https://doi.org/10.1177/074873099129000867* 
- 'awakeMu': Mu for awake state  
- 'awakeChi': Chi for awake state
- 'asleepMu': Mu for asleep state
- 'asleepChi': Chi for asleep state
- 'Irecep': Variable for calculation of rate of change in sleepModel
- 'targc': Variable for calculation of rate of change in sleepModel
- 'wakeThreshold': The point at which you wake up 
- 'sleepThreshold': The point at which you go to sleep
- 'sleepDriveSlope': The value of the slope of the sleep drive to be used in calculating the sleep drive model
- 'schoolStart': The start time of school (hours since midnight)
- 'schoolDuration': The time between when school begins and ends (hours)
- 'baselineLight': The amount of baseline light (lux)
- 'schoolBrightness': The light exposure in school (lux) 
- 'lightSlope': The slope of the light exposure curve (1/hours)
- 'lightRiseTime': The time of day when natural light exposure begins (hours)
- 'lightFallTime': The time of day when natural light exposure ends (hours)
- 'lightAmplitude': The amplitude of the light exposure curve (lux)
- 'lightBaseline': The baseline light exposure (lux)
}
\keyword{datasets}
