% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian.R
\name{getBHat}
\alias{getBHat}
\title{Calculate the BHat value based on n and light intensity}
\usage{
getBHat(n, light, params)
}
\arguments{
\item{n}{Numeric, the current value of n}

\item{light}{Numeric, the current light intensity}

\item{params}{List, a list of parameters including:
- G: the maximum growth rate
- alpha_0, P, I_0: parameters for the alpha calculation}
}
\value{
Numeric, the calculated BHat value
}
\description{
Calculate the BHat value based on n and light intensity
}
