% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{getR1b}
\alias{getR1b}
\title{Calculate the R1b value based on A, R1tot, and Psi}
\usage{
getR1b(A, R1tot)
}
\arguments{
\item{A}{Numeric, the value of A}

\item{R1tot}{Numeric, the total value of R1}
}
\value{
Numeric, the calculated R1b value
}
\description{
Calculate the R1b value based on A, R1tot, and Psi
}
