% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school.R
\name{isSchoolDay}
\alias{isSchoolDay}
\title{Check if a given time is within a school day}
\usage{
isSchoolDay(t)
}
\arguments{
\item{t}{Time in hours since the start of the week (0 = Sunday at midnight)}
}
\value{
TRUE if t is within a school day (Monday to Friday), FALSE otherwise
}
\description{
Check if a given time is within a school day
}
