% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school.R
\name{isStudentAwake}
\alias{isStudentAwake}
\title{Determine if a student is awake}
\usage{
isStudentAwake(t, R1bValue, Psi, inSchool, params)
}
\arguments{
\item{t}{Time in hours since the start of the week (0 = Sunday at midnight)}

\item{R1bValue}{The current value of the R1b variable in the circadian model}

\item{Psi}{The current value of the Psi variable in the circadian model}

\item{inSchool}{TRUE if the student is currently in school, FALSE otherwise}

\item{params}{A list containing 'sleepThreshold' and 'socialFactor'}
}
\value{
TRUE if the student is awake, FALSE otherwise
}
\description{
Determine if a student is awake
}
