% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school.R
\name{isWithinSchoolHours}
\alias{isWithinSchoolHours}
\title{Check if a given time is within school hours}
\usage{
isWithinSchoolHours(t, schoolStart, schoolDuration)
}
\arguments{
\item{t}{Time in hours since the start of the week (0 = Sunday at midnight)}

\item{schoolStart}{Start of the school day in hours since midnight}

\item{schoolDuration}{Duration of the school day in hours}
}
\value{
TRUE if t is within school hours, FALSE otherwise
}
\description{
Check if a given time is within school hours
}
