% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoprocess.R
\name{nextWakeState}
\alias{nextWakeState}
\title{Determine the next wake state of the student based on the current state}
\usage{
nextWakeState(t, S, R1b, Psi, inSchool, params)
}
\arguments{
\item{t}{Time in hours since the start of the week (0 = Sunday at midnight)}

\item{S}{The current wake state. 1 if awake, 0 if asleep}

\item{R1b}{The current value of the R1b variable in the circadian model}

\item{Psi}{The current value of the Psi variable in the circadian model}

\item{inSchool}{TRUE if the student is currently in school, FALSE otherwise}

\item{params}{A parameter list}
}
\value{
TRUE if the student is awake, FALSE otherwise
}
\description{
Determine the next wake state of the student based on the current state
}
\details{
Required parameters in `params` are:
- `sleepThreshold`: The sleep drive threshold for falling asleep
- `wakeThreshold`: The sleep drive threshold for waking up
}
