% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rootfunc}
\alias{rootfunc}
\title{Root function to detect when sleep or wake thresholds are crossed}
\usage{
rootfunc(t, x, params)
}
\arguments{
\item{t}{Time in hours since the start of the week (0 = Sunday at midnight)}

\item{x}{The current state of the system.}

\item{params}{The parameters of the system.}
}
\value{
Array with the distance to wake and sleep thresholds respectively.
}
\description{
Root function to detect when sleep or wake thresholds are crossed
}
