% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateCircadianModel}
\alias{simulateCircadianModel}
\title{Simulates the circadian model.}
\usage{
simulateCircadianModel(timeArray, lightArray, initialCondition, params)
}
\arguments{
\item{timeArray}{An array of time values.}

\item{lightArray}{An array of light values.}

\item{initialCondition}{The initial condition for the circadian model.}

\item{params}{A set of parameters for the circadian model.}
}
\value{
The output of the simulation.
}
\description{
This function takes in an array of time values, an array of light values,
an initial condition, and a set of parameters. It uses the circadian model
to simulate the circadian rhythm over time.
}
