% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateTwoProcessModel}
\alias{simulateTwoProcessModel}
\title{Simulate Two Process Model}
\usage{
simulateTwoProcessModel(timeArray, lightArray, initialCondition, params)
}
\arguments{
\item{timeArray}{A numeric vector representing the time points at which the model will be evaluated.}

\item{lightArray}{A numeric vector representing the light values corresponding to each time point.}

\item{initialCondition}{A numeric vector representing the initial condition of the model.}

\item{params}{A list of parameters required for the model simulation.}
}
\value{
A list containing the simulation results.
}
\description{
This function simulates a two-process model using the LSODE solver.
}
