% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{sleepParameters}
\alias{sleepParameters}
\title{Calculate the sleep parameters for the current wakefulness state}
\usage{
sleepParameters(t, isAwake, params)
}
\arguments{
\item{t}{Time in hours since the start of the week (0 = Sunday at midnight)}

\item{isAwake}{Logical, whether the student is awake.}

\item{params}{List, a list of parameters including:
\itemize{
\item awakeChi, asleepChi: Chi for awake and asleep states
\item awakeMu, asleepMu: Mu for awake and asleep states
}}
}
\value{
Numeric vector, the calculated mu, ichi, and light
}
\description{
Calculate the sleep parameters for the current wakefulness state
}
