% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school.R
\name{studentInSchool}
\alias{studentInSchool}
\title{Check if a student is in school at a given time}
\usage{
studentInSchool(t, params)
}
\arguments{
\item{t}{Time in hours since the start of the week (0 = Sunday at midnight)}

\item{params}{A list containing 'schoolStart' and 'schoolDuration'}
}
\value{
TRUE if a student is in school at time t, FALSE otherwise
}
\description{
Check if a student is in school at a given time
}
