% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light.R
\name{sunriseSunset}
\alias{sunriseSunset}
\title{Calculate sunrise and sunset times for a given location and date}
\usage{
sunriseSunset(date, latitude, longitude)
}
\arguments{
\item{date}{A date object}

\item{latitude}{Latitude of the location}

\item{longitude}{Longitude of the location}
}
\value{
A list with two elements: `sunrise` and `sunset`, each containing
the time of the respective event as a floating point number between 0 and 24
}
\description{
`sunriseSunset()` calculates the time of sunrise and sunset for a given
location and date. The function uses the `suncalc` package to calculate
the times based on the latitude and longitude of the location and the date.
}
