% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoprocess.R
\name{thresholdCrossingEvent}
\alias{thresholdCrossingEvent}
\title{Event function to detect when sleep or wake thresholds are crossed}
\usage{
thresholdCrossingEvent(t, x, light, params)
}
\arguments{
\item{t}{Time in hours since the start of the week (0 = Sunday at midnight)}

\item{x}{The current state of the system.}

\item{light}{The light input to the model.}

\item{params}{The parameters of the system.}
}
\value{
The state of the system with the wake state updated.
}
\description{
Event function to detect when sleep or wake thresholds are crossed
}
