% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoprocess.R
\name{twoProcessModel}
\alias{twoProcessModel}
\title{Model combining the circadian and sleep models}
\usage{
twoProcessModel(t, x, params)
}
\arguments{
\item{t}{Time in hours since the start of the week (0 = Sunday at midnight)}

\item{x}{Current values of the model variables: R, Psi, N, A, R1tot, and S.}

\item{params}{A list of parameters used in the model.}
}
\value{
Derivatives of the model variables: dR, dPsi, dN, dA, dR1tot, and dS.
}
\description{
Model combining the circadian and sleep models
}
