% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldCities.R
\docType{data}
\name{worldCities}
\alias{worldCities}
\title{Database containing world cities with their associated country and geonameid}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 26467 rows and 4 columns.
}
\source{
GeoNames
}
\usage{
worldCities
}
\description{
Database containing world cities with their associated country and geonameid
}
\details{
Data is taken from https://github.com/datasets/world-cities. We thank
Geonames, Lexman, and the Open Knowledge Foundation.
}
\keyword{datasets}
