% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoprocess.R
\name{distanceToThreshold}
\alias{distanceToThreshold}
\title{Function to calculate the distance to the wake and sleep thresholds}
\usage{
distanceToThreshold(t, x, light, params)
}
\arguments{
\item{t}{Time in hours since the start of the week (0 = Sunday at midnight)}

\item{x}{The current state of the system.}

\item{light}{The light input to the model.}

\item{params}{The parameters of the system.}
}
\value{
Array with the distance to wake and sleep thresholds respectively.
}
\description{
Function to calculate the distance to the wake and sleep thresholds
}
