% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian.R
\name{getAlpha}
\alias{getAlpha}
\title{Calculate the alpha value based on the current light intensity}
\usage{
getAlpha(light, params)
}
\arguments{
\item{light}{Numeric, the current light intensity}

\item{params}{List, a list of parameters including:
\itemize{
\item alpha_0: the maximum possible value of alpha
\item P: the exponent in the equation
\item I_0: the half-saturation constant
}}
}
\value{
Numeric, the calculated alpha value
}
\description{
Calculate the alpha value based on the current light intensity
}
