% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{sleepDrive}
\alias{sleepDrive}
\title{Calculate the sleep drive}
\usage{
sleepDrive(R1bValue, Psi, params)
}
\arguments{
\item{R1bValue}{The current value of the R1b variable in the circadian model}

\item{Psi}{The current value of the Psi variable in the circadian model}

\item{params}{A list containing 'sleepDriveSlope'}
}
\value{
The calculated sleep drive
}
\description{
Calculate the sleep drive
}
